Update tests to pass
